requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
cycler==0.11.0
Django==3.2.15
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-resized==1.0.1
django-taggit==3.0.0
django-tinymce==3.4.0 #modify from 2.6
fonttools==4.34.4
kiwisolver==1.4.4
matplotlib==3.5.1
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
psycopg2==2.9.3
gunicorn==20.1.0
